backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
Pillow==12.1.0
